+++ docs/changes.rst
@@ -12,6 +12,7 @@ Release history
 - feat: privacy API for workspaces: :attr:`api.workspace_settings.privacy <wxc_sdk.workspace_settings.WorkspaceSettingsApi.privacy>`
 - feat: barge API for workspaces: :attr:`api.workspace_settings.barge <wxc_sdk.workspace_settings.WorkspaceSettingsApi.barge>`
 - feat: new :meth:`api.workspace_settings.devices.list_and_counts <wxc_sdk.workspace_settings.devices.WorkspaceDevicesApi.list_and_counts>`
+- feat: barge API for virtual lines: :attr:`api.telephony.virtual_lines.barge <wxc_sdk.telephony.virtual_line.VirtualLinesApi.barge>`
 
 1.19.0
 ------